+++ com.peopleground.sagwim.fe/src/pages/admin/AdminPostListPage.tsx
@@ -11,13 +11,13 @@ import { LoadingSpinner } from '../../components/common/LoadingSpinner'
 import { Skeleton } from '../../components/common/Skeleton'
 import { ConfirmDialog } from '../../components/common/ConfirmDialog'
 import { SuccessDialog } from '../../components/common/SuccessDialog'
+import { Pagination } from '../../components/common/Pagination'
 import { formatDateTime } from '../../utils/dateUtils'
 import type { AdminContentResponse } from '../../types/post'
 import tableStyles from '../../components/admin/adminTable.module.css'
 import pageStyles from './AdminPostListPage.module.css'
 
 const PAGE_SIZE = 10
-const MAX_VISIBLE_PAGES = 5
 
 type ConfirmAction = 'delete' | 'restore'
 
@@ -100,19 +100,6 @@ export function AdminPostListPage() {
     } finally {
       setActionLoading(false)
     }
-  }
-
-  const getPageNumbers = (): number[] => {
-    if (totalPages <= MAX_VISIBLE_PAGES) {
-      return Array.from({ length: totalPages }, (_, i) => i)
-    }
-    const half = Math.floor(MAX_VISIBLE_PAGES / 2)
-    let start = Math.max(0, page - half)
-    const end = Math.min(totalPages, start + MAX_VISIBLE_PAGES)
-    if (end - start < MAX_VISIBLE_PAGES) {
-      start = Math.max(0, end - MAX_VISIBLE_PAGES)
-    }
-    return Array.from({ length: end - start }, (_, i) => start + i)
   }
 
   const isDeleted = (content: AdminContentResponse): boolean => content.deletedDate !== null
@@ -209,41 +196,12 @@ export function AdminPostListPage() {
               </table>
             </div>
 
-            {totalPages > 1 && (
-              <div className={tableStyles.paginationBar}>
-                <button
-                  type="button"
-                  className={tableStyles.pageButton}
-                  onClick={() => handlePageChange(page - 1)}
-                  disabled={loading || page === 0}
-                >
-                  이전
-                </button>
-                {getPageNumbers().map((pageNum) => (
-                  <button
-                    key={pageNum}
-                    type="button"
-                    className={
-                      pageNum === page
-                        ? tableStyles.pageButtonActive
-                        : tableStyles.pageButton
-                    }
-                    onClick={() => handlePageChange(pageNum)}
-                    disabled={loading}
-                  >
-                    {pageNum + 1}
-                  </button>
-                ))}
-                <button
-                  type="button"
-                  className={tableStyles.pageButton}
-                  onClick={() => handlePageChange(page + 1)}
-                  disabled={loading || page >= totalPages - 1}
-                >
-                  다음
-                </button>
-              </div>
-            )}
+            <Pagination
+              page={page}
+              totalPages={totalPages}
+              onPageChange={handlePageChange}
+              disabled={loading}
+            />
           </>
         )}
       </div>
